feat: New settings page
