+++ integration_test/tests/helpers.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:monekin/app/home/dashboard.page.dart';
 import 'package:monekin/app/onboarding/onboarding.dart';
-import 'package:monekin/app/settings/settings.page.dart';
+import 'package:monekin/app/settings/more_actions.page.dart';
 import 'package:monekin/core/database/services/app-data/app_data_service.dart';
 import 'package:monekin/core/database/services/user-setting/user_setting_service.dart';
 import 'package:monekin/i18n/generated/translations.g.dart';
@@ -45,5 +45,5 @@ Future<void> openMorePage(WidgetTester tester) async {
   await tester.tap(find.text(t.more.title));
   await tester.pumpAndSettle();
 
-  expect(find.byType(SettingsPage), findsOneWidget);
+  expect(find.byType(MoreActionsPage), findsOneWidget);
 }
