+++ lib/app/settings/widgets/transaction_swipe_action_selector.dart
@@ -45,7 +45,7 @@ class TransactionSwipeActionSelector extends StatelessWidget {
           action?.displayName(context) ?? null.displayName(context),
       valueGetter: (action) => action,
       title: title,
-      subtitle: t.settings.swipe_actions.choose_description,
+      subtitle: t.settings.transactions.swipe_actions.choose_description,
     );
   }
 }
