+++ lib/i18n/json/uk.json
@@ -747,52 +747,72 @@
     }
   },
   "SETTINGS": {
-    "title_long": "Налаштування та вигляд",
+    "title_long": "Налаштування та Персоналізація",
     "title_short": "Налаштування",
-    "description": "Тема додатку, тексти та інші загальні налаштування",
+    "description": "Тема, Мова, Дані та інше",
     "edit_profile": "Редагувати профіль",
-    "lang.section": "Мова та тексти",
-    "lang.title": "Мова додатку",
-    "lang.descr": "Мова, в якій будуть відображатися тексти в додатку",
-    "lang.help": "Якщо ви хочете співпрацювати з перекладами цієї програми, ви можете звернутися до <a href='https://github.com/enrique-lozano/Monekin/tree/main/lib/i18n'>нашого посібник</ a>",
-    "locale": "Регіон",
-    "locale_auto": "Система",
-    "locale.descr": "Встановіть формат, який буде використовуватися для дат, чисел...",
-    "locale.warn": "Після зміни регіону додаток оновиться",
-    "first_day_of_week": "Перший день тижня",
-    "theme_and_colors": "Тема та кольори",
-    "theme": "Тема",
-    "theme.auto": "система",
-    "theme.light": "Світла",
-    "theme.dark": "Темна",
-    "amoled-mode": "Режим AMOLED",
-    "amoled-mode.descr": "Використовуйте чисто чорний шпалери, якщо це можливо. Це трохи допоможе акумулятору пристроїв з екранами AMOLED",
-    "dynamic-colors": "Динамічні кольори",
-    "dynamic-colors.descr": "Використовуйте колір акценту вашої системи, коли це можливо",
-    "accent-color": "Колір акценту",
-    "accent-color.descr": "Виберіть колір, який додаток буде використовувати для виділення певних частин інтерфейсу",
-    "font": "Шрифт",
-    "font-platform": "Платформа",
-    "SWIPE-ACTIONS": {
-      "title": "Дії гортання",
-      "choose-description": "Виберіть, яка дія буде ініційована, коли ви проводите пальцем по елементу транзакції у списку в цьому напрямку",
-      "swipe-left": "Проведіть ліворуч",
-      "swipe-right": "Проведіть праворуч",
-      "none": "Ніяких дій",
-      "toggle-reconciled": "Перемикач узгоджено",
-      "toggle-pending": "Перемкнути в очікуванні",
-      "toggle-voided": "Перемикач скасовано",
-      "toggle-unreconciled": "Перемикач неузгоджений",
-      "remove-status": "Видалити статус"
+    "GENERAL": {
+      "menu_title": "Загальні налаштування",
+      "menu_descr": "Мова, конфіденційність та інше",
+      "LANGUAGE": {
+        "section": "Мова та тексти",
+        "title": "Мова додатку",
+        "descr": "Мова, в якій будуть відображатися тексти в додатку",
+        "help": "Якщо ви хочете співпрацювати з перекладами цієї програми, ви можете звернутися до <a href='https://github.com/enrique-lozano/Monekin/tree/main/lib/i18n'>нашого посібник</ a>"
+      },
+      "LOCALE": {
+        "title": "Регіон",
+        "auto": "Система",
+        "descr": "Встановіть формат, який буде використовуватися для дат, чисел...",
+        "warn": "Після зміни регіону додаток оновиться",
+        "first_day_of_week": "Перший день тижня"
+      }
     },
     "SECURITY": {
       "title": "Безпека",
-      "private-mode-at-launch": "Приватний режим під час запуску",
-      "private-mode-at-launch.descr": "За замовчуванням запускати програму в приватному режимі",
-      "private-mode": "Приватний режим",
-      "private-mode.descr": "Приховати всі грошові значення",
-      "private-mode.activated": "Приватний режим активовано",
-      "private-mode.deactivated": "Приватний режим вимкнено"
+      "private_mode_at_launch": "Приватний режим під час запуску",
+      "private_mode_at_launch_descr": "За замовчуванням запускати програму в приватному режимі",
+      "private_mode": "Приватний режим",
+      "private_mode_descr": "Приховати всі грошові значення",
+      "private_mode_activated": "Приватний режим активовано",
+      "private_mode_deactivated": "Приватний режим вимкнено"
+    },
+    "TRANSACTIONS": {
+      "menu_title": "Транзакції",
+      "menu_descr": "Налаштуйте поведінку ваших транзакцій",
+      "title": "Налаштування транзакцій",
+      "SWIPE_ACTIONS": {
+        "title": "Дії гортання",
+        "choose_description": "Виберіть, яка дія буде ініційована, коли ви проводите пальцем по елементу транзакції у списку в цьому напрямку",
+        "swipe-left": "Проведіть ліворуч",
+        "swipe-right": "Проведіть праворуч",
+        "none": "Ніяких дій",
+        "toggle-reconciled": "Перемикач узгоджено",
+        "toggle-pending": "Перемкнути в очікуванні",
+        "toggle-voided": "Перемикач скасовано",
+        "toggle-unreconciled": "Перемикач неузгоджений",
+        "remove-status": "Видалити статус"
+      }
+    },
+    "APPEARANCE": {
+      "menu_title": "Тема та стиль",
+      "menu_descr": "Вибір теми, кольори та інші речі, пов'язані з виглядом програми",
+      "theme_and_colors": "Тема та кольори",
+      "THEME": {
+        "title": "Тема",
+        "auto": "система",
+        "light": "Світла",
+        "dark": "Темна"
+      },
+      "amoled_mode": "Режим AMOLED",
+      "amoled_mode_descr": "Використовуйте чисто чорний шпалери, якщо це можливо. Це трохи допоможе акумулятору пристроїв з екранами AMOLED",
+      "dynamic_colors": "Динамічні кольори",
+      "dynamic_colors_descr": "Використовуйте колір акценту вашої системи, коли це можливо",
+      "accent_color": "Колір акценту",
+      "accent_color_descr": "Виберіть колір, який додаток буде використовувати для виділення певних частин інтерфейсу",
+      "text": "Текст",
+      "font": "Шрифт",
+      "font_platform": "Платформа"
     }
   },
   "MORE": {
